Fix procedure for applying extra patches
In the 'lenny' kernel package, the 'setup' rule copies source
directories to build directories so it is then too late to apply
patches in a source directory.  There is no need to run the rule
explicitly, anyway.

Also, the explanation of which directory to apply patches in is
a bit ambiguous, so make that clearer.