Move the web site to http://kernel-handbook.debian.net
Unfortunately Salsa is not ready to host static sites, and may not be
ready until after Alioth is shut down.  For now, use a debian.net
domain hosted on my own web server.  This can be redirected to Salsa
when it's ready.

My web server has MultiViews enabled by default and doesn't allow
this to be set in .htaccess files.