Remove obsolete text referring to "patchlevels" in source packages
Before 2011, kernel udebs were built from separate source packages,
and dak didn't support Built-Using.  This meant there was no guarantee
that the exact source version for the kernel udebs in a suite would be
present in the same suite.

To ensure that we could comply with the GPL's requirements for source
availability, the kernel patch system supported multiple series files
that could each add and remove patch files.  Each Debian revision
based on the same upstream version would add a new series file, and
the source could be patched to the current or an older "patchlevel".

This became obsolete once we started building kernel udebs directly,
and the patch system has since been entirely replaced with quilt.