Remove redundant "bash" from debian/bin/test-patches command lines
When test-patches was introduced, src:linux was still using source
format 1.0 which did not support setting executable bits on files in
the debian/ directory (except debian/rules).  Thus it was necessary to
include the script interpreter in these command lines.

This hasn't been the case for many years, and putting "bash" in front
will cause problems if we ever want to change the script's
interpreter.