Deprecate older versions of test-patches
The current version has important bugs and limitations that make it
not nearly as easy to use as it should be.  This is now fixed in
unstable and will soon be fixed for bookworm, but older releases may
remain broken.  Expand the warning about debug info to add 2 other big
problems.