Recommend changing ABI name before rebuilding official kernel packages
When building kernel packages with experimental changes, it is
undesirable that these packages will replace or conflict with the
current kernel packages.  Aside from any run-time bugs, if Secure Boot
is still enabled then the unsigned linux-image package will not be
bootable.  Recommend changing the ABI name so the new binary packages
are co-installable.

Give straightforward instructions for changing the ABI name for
a custom package, and given an example name of "0.local".